requirements.txt
asgiref==3.11.0
Django==6.0
django-admin-interface==0.32.0
django-colorfield==0.14.0
django-tinymce==5.0.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
python-slugify==8.0.4
sqlparse==0.5.4
text-unidecode==1.3
tzdata==2025.2
